rename 'server_client' to 'server_web_client'
